requirements.txt
annotated-types==0.7.0
anyio==4.12.0
bitsandbytes==0.41.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
distro==1.9.0
duckduckgo_search==6.1.0
exceptiongroup==1.3.1
filelock==3.20.0
fonttools==4.61.1
fsspec==2025.12.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
intel-openmp==2021.4.0
isal==1.8.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
kiwisolver==1.4.9
langdetect==1.0.9
MarkupSafe==3.0.2
matplotlib==3.10.8
mkl==2021.4.0
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
openai==2.12.0
orjson==3.11.5
orjsonl==1.0.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.1.4
pyreqwest_impersonate==0.5.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.5.1
scipy==1.12.0
sentence-transformers==5.2.0
six==1.17.0
sniffio==1.3.1
sty==1.0.6
sympy==1.14.0
tbb==2021.13.1
threadpoolctl==3.6.0
tiktoken==0.7.0
tokenizers==0.22.1
torch==2.11.0.dev20251216+cu128
torchaudio==2.10.0.dev20251216+cu128
torchvision==0.25.0.dev20251216+cu128
tqdm==4.67.1
transformers==4.57.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
xopen==2.0.2
zlib-ng==1.0.0
